all auths go to (file) heaven
