+++ plugins/module_utils/_authenticator.py
@@ -13,16 +13,14 @@
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
+# please keep this list in alphabetical order of auth method name
+from ansible_collections.community.hashi_vault.plugins.module_utils._auth_method_approle import HashiVaultAuthMethodApprole
+from ansible_collections.community.hashi_vault.plugins.module_utils._auth_method_aws_iam_login import HashiVaultAuthMethodAwsIamLogin
+from ansible_collections.community.hashi_vault.plugins.module_utils._auth_method_jwt import HashiVaultAuthMethodJwt
+from ansible_collections.community.hashi_vault.plugins.module_utils._auth_method_ldap import HashiVaultAuthMethodLdap
 from ansible_collections.community.hashi_vault.plugins.module_utils._auth_method_none import HashiVaultAuthMethodNone
-
-from ansible_collections.community.hashi_vault.plugins.module_utils._hashi_vault_common import (
-    HashiVaultAuthMethodToken,
-    HashiVaultAuthMethodUserpass,
-    HashiVaultAuthMethodAwsIamLogin,
-    HashiVaultAuthMethodLdap,
-    HashiVaultAuthMethodApprole,
-    HashiVaultAuthMethodJwt,
-)
+from ansible_collections.community.hashi_vault.plugins.module_utils._auth_method_token import HashiVaultAuthMethodToken
+from ansible_collections.community.hashi_vault.plugins.module_utils._auth_method_userpass import HashiVaultAuthMethodUserpass
 
 
 class HashiVaultAuthenticator():
